+++ includes/shortcodes/renderers/call_to_action.php
@@ -5,9 +5,9 @@
  * This file is being included into "../class-shortcodes.php" file's shortcode_render() method.
  *
  * @since    1.0
- * @version  1.2
+ * @version  1.2.9.1
  *
- * @uses   $codes_globals['colors'], $codes_globals['sizes']['values']
+ * @uses  $codes_globals['colors'], $codes_globals['sizes']['values']
  *
  * @param  string button_class
  * @param  string button_color
@@ -17,7 +17,7 @@
  * @param  string button_url
  * @param  string caption
  * @param  string class
- * @param  string heading_tag (heading tag setup option for better SEO)
+ * @param  string heading_tag (heading tag option for better accessibility setup)
  * @param  string ... You can actually set up a custom attributes for this shortcode. They will be outputed as HTML attributes for the button.
  */
 
@@ -48,7 +48,7 @@
 		$atts['content'] = apply_filters( 'wmhook_shortcode_' . $shortcode . '_content', $atts['content'], $atts );
 	//button_color
 		$atts['button_color'] = trim( $atts['button_color'] );
-		if ( in_array( $atts['button_color'], array_keys( $codes_globals['colors'] ) ) ) {
+		if ( $atts['button_color'] && in_array( $atts['button_color'], array_keys( $codes_globals['colors'] ) ) ) {
 			$atts['button_class'] .= ' color-' . $atts['button_color'];
 			$atts['class']        .= ' cta-button-color-' . $atts['button_color'];
 		}
@@ -59,7 +59,7 @@
 		}
 	//button_size
 		$atts['button_size'] = trim( $atts['button_size'] );
-		if ( in_array( $atts['button_size'], array_keys( $codes_globals['sizes']['values'] ) ) ) {
+		if ( $atts['button_size'] && in_array( $atts['button_size'], array_keys( $codes_globals['sizes']['values'] ) ) ) {
 			$atts['button_class'] .= ' size-' . $codes_globals['sizes']['values'][ $atts['button_size'] ];
 		}
 	//button_text