+++ includes/shortcodes/renderers/dropcap.php
@@ -5,9 +5,9 @@
  * This file is being included into "../class-shortcodes.php" file's shortcode_render() method.
  *
  * @since    1.0
- * @version  1.1.6
+ * @version  1.2.9.1
  *
- * @uses   $codes_globals['colors'], $codes_globals['dropcap_shapes']
+ * @uses  $codes_globals['colors'], $codes_globals['dropcap_shapes']
  *
  * @param  string class
  * @param  string color
@@ -33,12 +33,12 @@
 		$atts['class'] = trim( 'wm-dropcap ' . trim( $atts['class'] ) );
 	//color
 		$atts['color'] = trim( $atts['color'] );
-		if ( in_array( $atts['color'], array_keys( $codes_globals['colors'] ) ) ) {
+		if ( $atts['color'] && in_array( $atts['color'], array_keys( $codes_globals['colors'] ) ) ) {
 			$atts['class'] .= ' color-' . $atts['color'];
 		}
 	//shape
 		$atts['shape'] = trim( $atts['shape'] );
-		if ( in_array( $atts['shape'], array_keys( $codes_globals['dropcap_shapes'] ) ) ) {
+		if ( $atts['shape'] && in_array( $atts['shape'], array_keys( $codes_globals['dropcap_shapes'] ) ) ) {
 			$atts['class'] .= ' shape-' . $atts['shape'];
 		}
 	//class