+++ includes/shortcodes/renderers/icon.php
@@ -5,16 +5,16 @@
  * This file is being included into "../class-shortcodes.php" file's shortcode_render() method.
  *
  * @since    1.0
- * @version  1.1.6
+ * @version  1.2.9.1
  *
- * @uses     $codes_globals['sizes']['values']
+ * @uses  $codes_globals['sizes']['values']
  *
- * @param    string class
- * @param    string size
- * @param    string social
- * @param    string style
- * @param    string url
- * @param    string ... You can actually set up a custom attributes for this shortcode. They will be outputed as HTML attributes.
+ * @param  string class
+ * @param  string size
+ * @param  string social
+ * @param  string style
+ * @param  string url
+ * @param  string ... You can actually set up a custom attributes for this shortcode. They will be outputed as HTML attributes.
  */
 
 
@@ -50,7 +50,7 @@
 		$atts['url'] = ( $atts['social'] ) ? ( esc_url( $atts['url'] ) ) : ( '' );
 	//size
 		$atts['size'] = trim( $atts['size'] );
-		if ( in_array( $atts['size'], array_keys( $codes_globals['sizes']['values'] ) ) ) {
+		if ( $atts['size'] && in_array( $atts['size'], array_keys( $codes_globals['sizes']['values'] ) ) ) {
 			$atts['class'] .= ' size-' . $codes_globals['sizes']['values'][ $atts['size'] ];
 		}
 	//style