+++ includes/shortcodes/renderers/message.php
@@ -5,13 +5,13 @@
  * This file is being included into "../class-shortcodes.php" file's shortcode_render() method.
  *
  * @since    1.0
- * @version  1.2
+ * @version  1.2.9.1
  *
- * @uses   $this->$codes_globals['colors'], $codes_globals['sizes']['values']
+ * @uses  $this->$codes_globals['colors'], $codes_globals['sizes']['values']
  *
  * @param  string class
  * @param  string color
- * @param  string heading_tag (heading tag setup option for better SEO)
+ * @param  string heading_tag (heading tag option for better accessibility setup)
  * @param  string icon
  * @param  string size
  * @param  string title
@@ -41,7 +41,7 @@
 		$atts['content'] = '<div class="wm-message-content wm-message-element' . esc_attr( $iconated ) . '">' . $atts['content'] . '</div>';
 	//color
 		$atts['color'] = trim( $atts['color'] );
-		if ( in_array( $atts['color'], array_keys( $codes_globals['colors'] ) ) ) {
+		if ( $atts['color'] && in_array( $atts['color'], array_keys( $codes_globals['colors'] ) ) ) {
 			$atts['class'] .= ' color-' . $atts['color'];
 		}
 	//icon
@@ -51,7 +51,7 @@
 		}
 	//size
 		$atts['size'] = trim( $atts['size'] );
-		if ( in_array( $atts['size'], array_keys( $codes_globals['sizes']['values'] ) ) ) {
+		if ( $atts['size'] && in_array( $atts['size'], array_keys( $codes_globals['sizes']['values'] ) ) ) {
 			$atts['class'] .= ' size-' . $codes_globals['sizes']['values'][ $atts['size'] ];
 		}
 	//title