+++ includes/shortcodes/renderers/price.php
@@ -5,13 +5,13 @@
  * This file is being included into "../class-shortcodes.php" file's shortcode_render() method.
  *
  * @since    1.0
- * @version  1.2
+ * @version  1.2.9.1
  *
  * @param  string caption
  * @param  string class
  * @param  string color
  * @param  string cost
- * @param  string heading_tag (heading tag setup option for better SEO)
+ * @param  string heading_tag (heading tag option for better accessibility setup)
  * @param  string type Legacy attribute
  * @param  string appearance Introduced not to conflict with Beaver Builder
  */