+++ includes/shortcodes/renderers/progress.php
@@ -5,9 +5,9 @@
  * This file is being included into "../class-shortcodes.php" file's shortcode_render() method.
  *
  * @since    1.0
- * @version  1.1.6
+ * @version  1.2.9.1
  *
- * @uses   $codes_globals['colors']
+ * @uses  $codes_globals['colors']
  *
  * @param  string class
  * @param  string color
@@ -32,7 +32,7 @@
 		$atts['class_bar']     = trim( 'wm-progress-bar' );
 	//color
 		$atts['color'] = trim( $atts['color'] );
-		if ( in_array( $atts['color'], array_keys( $codes_globals['colors'] ) ) ) {
+		if ( $atts['color'] && in_array( $atts['color'], array_keys( $codes_globals['colors'] ) ) ) {
 			$atts['class']     .= ' container-color-' . $atts['color'];
 			$atts['class_bar'] .= ' color-' . $atts['color'];
 		}